+++ src/edu/saddleback/cs4b/Backend/Messages/ActiveUserMessage.java
@@ -1,31 +1,9 @@
 package edu.saddleback.cs4b.Backend.Messages;
 
-import java.util.List;
-
-
-public class ActiveUserMessage extends BaseMessage {
-    private List<String> activeUsers;
-
+public class ActiveUserMessage extends BaseMessage
+{
     public ActiveUserMessage()
     {
-        this(null);
+        super(MsgTypes.ACTIVE_USER_RESPONSE);
     }
-
-    public ActiveUserMessage(List<String> newActiveUsers)
-    {
-        super(MsgTypes.ACTIVE_USER);
-        setActiveUsers(newActiveUsers);
-    }
-
-    public void setActiveUsers(List<String> newActiveUsers)
-    {
-        activeUsers = newActiveUsers;
-    }
-
-    public List<String> getActiveUsers()
-    {
-        return activeUsers;
-    }
-
-
 }
