+++ src/edu/saddleback/cs4b/Backend/Messages/ActiveUserResponseMessage.java
@@ -2,14 +2,30 @@ package edu.saddleback.cs4b.Backend.Messages;
 
 import java.util.List;
 
-public class ActiveUserResponseMessage extends BaseMessage
-{
+
+public class ActiveUserResponseMessage extends BaseMessage {
+    private List<String> activeUsers;
 
     public ActiveUserResponseMessage()
     {
-        super(MsgTypes.ACTIVE_USER_RESPONSE);
+        this(null);
     }
 
+    public ActiveUserResponseMessage(List<String> newActiveUsers)
+    {
+        super(MsgTypes.ACTIVE_USER);
+        setActiveUsers(newActiveUsers);
+    }
+
+    public void setActiveUsers(List<String> newActiveUsers)
+    {
+        activeUsers = newActiveUsers;
+    }
+
+    public List<String> getActiveUsers()
+    {
+        return activeUsers;
+    }
 
 
 }
